Fix error in MMC1 implementation.
1. Should or in 0x0C, not mask it out.

This error was triggered by a bug in Z2, where there is an errant write
from the Daira axe-throwing routine to $95e6.  It would but the PPU into
mirrored-0 mode and make the screen look like it was in "scroll lock"
mode, as triggered by the healer glitch (it really wasn't).